add Test to all components
